+++ leetcode-tui-db/src/models/question.rs
@@ -20,6 +20,20 @@ pub struct DbQuestion {
     pub topics: Vec<DbTopic>,
 }
 
+impl Ord for DbQuestion {
+    fn cmp(&self, other: &Self) -> std::cmp::Ordering {
+        self.id.cmp(&other.id)
+    }
+}
+
+impl PartialOrd for DbQuestion {
+    fn partial_cmp(&self, other: &Self) -> Option<std::cmp::Ordering> {
+        self.id.partial_cmp(&other.id)
+    }
+}
+
+impl Eq for DbQuestion {}
+
 impl Display for DbQuestion {
     fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
         let mut w = String::new();
@@ -120,8 +134,10 @@ impl DbQuestion {
     fn update_in_db<'a>(&self) -> DBResult<Vec<Self>> {
         let rw = get_db_client().rw_transaction()?;
         let old = Self::get_question_by_id(self.id)?;
-        rw.update(old, self.clone())?;
-        rw.commit()?;
+        if let Some(old_q) = old {
+            rw.update(old_q, self.clone())?;
+            rw.commit()?;
+        }
         Ok(vec![self.clone()])
     }
 
@@ -131,12 +147,9 @@ impl DbQuestion {
         Ok(x.all().count())
     }
 
-    pub fn get_question_by_id<'a>(id: u32) -> DBResult<Self> {
+    pub fn get_question_by_id<'a>(id: u32) -> DBResult<Option<Self>> {
         let r = get_db_client().r_transaction()?;
-        let x = r
-            .get()
-            .primary::<DbQuestion>(id)?
-            .ok_or(DbErr::QuestionsNotFoundInDb(id.to_string()))?;
+        let x = r.get().primary::<DbQuestion>(id)?;
         // x.topics = x.fetch_all_topics(db)?;
         Ok(x)
     }
@@ -145,18 +158,47 @@ impl DbQuestion {
         save_multiple(&self.topics)
     }
 
-    // pub fn fetch_all_topics<'a>(&self, ) -> DBResult<Vec<DbTopic>> {
-    //     let q_topic_map = QuestionTopicMap::get_all_topic_slug_by_question(self, db)?;
-    //     let mut topics = vec![];
-
-    //     for topic_slug in q_topic_map {
-    //         topics.push(DbTopic::get_topic_by_slug(topic_slug.as_str(), db)?);
-    //     }
-    //     Ok(topics)
-    // }
-
     pub(crate) fn get_topics(&self) -> &Vec<DbTopic> {
         &self.topics
+    }
+
+    fn get_topic_question_mapping(&self) -> Vec<TopicQuestionMap> {
+        self.get_topics()
+            .iter()
+            .map(|q| TopicQuestionMap::new(&q.slug, self.id))
+            .collect::<Vec<_>>()
+    }
+
+    fn get_question_topic_mapping(&self) -> Vec<QuestionTopicMap> {
+        self.get_topics()
+            .iter()
+            .map(|q| QuestionTopicMap::new(self.id, &q.slug))
+            .collect::<Vec<_>>()
+    }
+
+    pub fn save_multiple_to_db(questions: Vec<Self>) {
+        let topic_question_map = questions
+            .iter()
+            .map(|q| q.get_topic_question_mapping())
+            .flatten()
+            .collect::<Vec<_>>();
+
+        let question_topic_map = questions
+            .iter()
+            .map(|q| q.get_question_topic_mapping())
+            .flatten()
+            .collect::<Vec<_>>();
+
+        let topics = questions
+            .iter()
+            .map(|q| q.get_topics().iter().map(|t| t.clone()))
+            .flatten()
+            .collect::<Vec<_>>();
+
+        save_multiple(&topic_question_map).unwrap();
+        save_multiple(&question_topic_map).unwrap();
+        save_multiple(&topics).unwrap();
+        save_multiple(&questions).unwrap();
     }
 
     pub fn save_to_db<'a>(&mut self) -> DBResult<bool> {